fix: massive edit typos. Classes logic fix, init images loading
