+++ components/PopupWithImage.js
@@ -1,15 +1,14 @@
 import  Popup  from './Popup.js';
 export default class PopupWithImage extends Popup {
     constructor(popupSelector){
-        super(popupSelector)
+        super(popupSelector);
+        this._image = this._element.querySelector("popup-image");
+        this._title = this._element.querySelector("popup__title");
     }
     open (data){
         super.open();
-        const placeImage = this._querySelector("popup-image");
-        const placeName = this._querySelector("popup__title");
-        placeImage.scr = data.link;
-        placeName.textContent = data.name;
-        placeImage.alt = data.name;
-
+        this._image.src = data.link;
+        this._title.textContent = data.name;
+        this._image.alt = data.name;
     }
 }
