+++ components/Section.js
@@ -1,4 +1,4 @@
-export default  class Sectuion{
+export default  class Section{
     constructor({ items, renderer }, containerSelector) {
         this._initialArray = items;
         this._renderer = renderer;
