+++ index.html
@@ -82,7 +82,7 @@
       </div>
     </article>
   </template>
-    <script type="module" src="./scripts/FormValidator.js"></script>
-    <script type="module" src="./scripts/Index.js"></script>
+    <!-- <script type="module" src="./scripts/FormValidator.js"></script> -->
+    <script type="module" src="./pages/index.js"></script>
   </body>
 </html>
