+++ pages/Index.js
@@ -1,39 +1,39 @@
 import { Card } from '../components/Card.js';
 import { initialCards } from '../components/initialCards.js';
 import { FormValidator } from '../components/FormValidator.js';
-import { selectors }  from '../components/selectors.js';
-//import Sectuion from '../components/Section.js';
-//import PopupWithForm from '../components/PopupWithForm.js';
-//import PopupWithImage from '../components/PopupWithImage';
-//import UserInfo from '../components/UserInfo.js';
+import { selectors } from '../components/selectors.js';
+import PopupWithImage from '../components/PopupWithImage.js';
+// import Section from '../components/Section.js';
+// import PopupWithForm from '../components/PopupWithForm.js';
+// import UserInfo from '../components/UserInfo.js';
 
 import {
   editFormElement,
-    nameInput,
-    jobInput,
-    editButton,
-    popupEdit,
-    profileTitle,
-    profileSubtitle,
-    elements,
-    popupAdd,
-    addButton,
-    addFormElement,
-    titleInput, 
-     linkInput ,
-     popupImage,
-     popupImageImage, 
-     popupImageCaption, 
-     popupAddForm,
-     popupEditForm ,
-     popups, 
-} from './utils/constant.js';
+  nameInput,
+  jobInput,
+  editButton,
+  popupEdit,
+  profileTitle,
+  profileSubtitle,
+  elements,
+  popupAdd,
+  addButton,
+  addFormElement,
+  titleInput,
+  linkInput,
+  popupImage,
+  popupImageImage,
+  popupImageCaption,
+  popupAddForm,
+  popupEditForm,
+  popups,
+} from '../utils/constant.js';
 
-const popupWImage = new PopupWithImage(imageInPopup);
+const popupWImage = new PopupWithImage(popupImage);
 popupWImage.setEventListeners();
 
-const popupWithDeleteCard = new PopupWithDelete({ popupSelector: deleteCard });
-popupWithDeleteCard.setEventListeners();
+// const popupWithDeleteCard = new PopupWithDelete({ popupSelector: deleteCard });
+// popupWithDeleteCard.setEventListeners();
 
 
 const closeByEsc = (evt) => {
