--- pages/utils/constant.js
@@ -1,41 +1,0 @@
-const editFormElement = document.querySelector('.popup-edit__form');
-const nameInput = document.querySelector('.popup__input_type_name');
-const jobInput = document.querySelector('.popup__input_type_about');
-const editButton = document.querySelector('.profile__edit-button');
-const popupEdit = document.querySelector('.popup-edit');
-const profileTitle = document.querySelector('.profile__title');
-const profileSubtitle = document.querySelector('.profile__subtitle');
-const elements = document.querySelector('.elements');
-const popupAdd = document.querySelector('.popup-add');
-const addButton = document.querySelector('.profile__add-button');
-const addFormElement = document.querySelector('.popup-add__form');
-const titleInput = document.querySelector('.popup__input_type_title');
-const linkInput = document.querySelector('.popup__input_type_link');
-const popupImage = document.querySelector('.popup-image');
-const popupImageImage = popupImage.querySelector('.popup-image__image');
-const popupImageCaption = popupImage.querySelector('.popup-image__caption');
-const popupAddForm = document.querySelector('.popup-add__form');
-const popupEditForm = document.querySelector('.popup-edit__form');
-const popups = document.querySelectorAll('.popup');
-
-export{
-    editFormElement,
-    nameInput,
-    jobInput,
-    editButton,
-    popupEdit,
-    profileTitle,
-    profileSubtitle,
-    elements,
-    popupAdd,
-    addButton,
-    addFormElement,
-    titleInput, 
-     linkInput ,
-     popupImage,
-     popupImageImage, 
-     popupImageCaption, 
-     popupAddForm,
-     popupEditForm ,
-     popups, 
-};
